FEATURE: padd-airports-to-table now includes skyscannerid query. Progress bar also added
