+++ utils/code_analysis_utils/get-all-functions-in-dir/go.mod
@@ -1,0 +1,3 @@
+module get-all-functions-in-dir
+
+go 1.18
