--- utils/database_utils/flights/add-flights-to-table/aerodatabox-sql/go.mod
@@ -1,8 +1,0 @@
-module aerodatabox-sql
-
-go 1.18
-
-require (
-	github.com/mattn/go-sqlite3 v1.14.22 // indirect
-	gopkg.in/yaml.v2 v2.4.0 // indirect
-)
