+++ utils/database_utils/flights/update-schedule/aerodatabox/.gitignore
@@ -1,0 +1,4 @@
+*.db
+*.exe
+aerodatabox-sql
+
