+++ utils/database_utils/flights/update-schedule/aerodatabox/fetch-these-flights.yaml
@@ -1,0 +1,12 @@
+flights:
+  - direction: "Departure"
+    airport: "BER"
+    startDate: "21-04-2024"
+    endDate: "22-04-2024"
+  - direction: "Arrival"
+    airport: "BER"
+    startDate: "23-04-2024"
+    endDate: "25-04-2024"
+ 
+      # You can add more entries with different airports, directions, and date ranges.
+
