+++ utils/database_utils/flights/update-schedule/aerodatabox/go.mod
@@ -1,0 +1,13 @@
+module aerodatabox
+
+go 1.18
+
+require (
+	github.com/mattn/go-sqlite3 v1.14.22 // indirect
+	github.com/mitchellh/colorstring v0.0.0-20190213212951-d06e56a500db // indirect
+	github.com/rivo/uniseg v0.4.7 // indirect
+	github.com/schollz/progressbar/v3 v3.14.2 // indirect
+	golang.org/x/sys v0.17.0 // indirect
+	golang.org/x/term v0.17.0 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
+)
