--- utils/get-all-functions-in-dir/go.mod
@@ -1,3 +1,0 @@
-module get-all-functions-in-dir
-
-go 1.18
